Modified demo_image.py to accept video as input.
Primary changes are running cv2.VideoCapture in a loop (instead of 
loading single image) and saving to an .mp4 file. Speed of analyzing a 
video can be increased by modifying input parameters --process_speed (to 
change at how many different resolutions the model analyzes each frame) 
and --frame_ratio which analyzes every [n] frames and saves the video 
accordingly (e.g. analyzes one out of every 2 frames in a 30 fps video, 
then saves it as a 15 fps so the timing of action is still correct).